WIP: pushing up for access on desktop. working on getting either the word to re-render or the entire wordbox on space when the next word would be queued. i seem to be getting close (hopefully)
